server/pyproject.toml
[project]
name = "goalmap-server"
version = "0.1.0"
description = "FastAPI backend for GoalMap AI using LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "requests>=2.32.0",
    "pydantic-settings>=2.12.0",
    "langgraph-checkpoint-postgres>=3.0.3",
    "psycopg[binary]>=3.3.2",
    "asyncpg>=0.31.0",
    "langfuse-langchain>=2.60.10.1",
    "langchain>=1.2.7",
    "pyjwt>=2.10.1",
    "cryptography>=42.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
